+++ src/collective/contact/core/indexers.py
@@ -48,7 +48,15 @@ def position_searchable_text(obj):
 @indexer(IPerson)
 def person_searchable_text(obj):
     results = []
-    results.append(safe_unicode(obj.SearchableText()))
+    use_description = api.portal.get_registry_record(
+        "collective.contact.core.interfaces.IContactCoreParameters."\
+        "use_description_to_search_person")
+    if use_description:
+        text = obj.SearchableText()
+    else:
+        text = obj.Title()
+
+    results.append(safe_unicode(text))
     use_held_positions = api.portal.get_registry_record(
         "collective.contact.core.interfaces.IContactCoreParameters."\
         "use_held_positions_to_search_person")